why there's nothing
